requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
tailwindcss
jinja2
sqlalchemy
psycopg2
